+++ lib/furikake/resources/ebs.rb
@@ -15,12 +15,16 @@ module Furikake
             }
           ]
         }
+        if $output_tag_keys and $output_tag_keys.length > 0
+          contents[:resources][0][:header] << 'Tags'
+        end
         Furikake::Formatter.shaping(format, contents)
       end
 
       def get_resources
         ec2 = Aws::EC2::Client.new
         params = {}
+        params[:filters] = $filters
         volumes = []
         loop do
           res = ec2.describe_volumes(params)
@@ -46,6 +50,16 @@ module Furikake
               attachments << attachment.join(':')
             end
             volume << attachments.join('<br>')
+
+            if $output_tag_keys
+              output_tags = []
+              $output_tag_keys.each do |t|
+                v.tags.each do |tag|
+                  output_tags << '"' + t + '":"' + tag.value + '"' if tag.key == t
+                end
+              end
+              volume << output_tags.sort.join('<br>')
+            end
 
             volumes << volume
           end