+++ lib/furikake/resources/redshift.rb
@@ -14,6 +14,9 @@ module Furikake
               }
           ]
         }
+        if $output_tag_keys and $output_tag_keys.length > 0
+          contents[:resources][0][:header] << 'Tags'
+        end
         Furikake::Formatter.shaping(format, contents).chomp
       end
 
@@ -42,6 +45,16 @@ module Furikake
             security_groups << (c[:vpc_security_groups].map {|c| c[:vpc_security_group_id]})
           end
           cluster << security_groups.sort.join('<br>')
+
+          if $output_tag_keys
+            output_tags = []
+            $output_tag_keys.each do |t|
+              c[:tags].each do |tag|
+                output_tags << '"' + t + '":"' + tag[:value] + '"' if tag[:key] == t
+              end
+            end
+            cluster << output_tags.sort.join('<br>')
+          end
 
           cluster_infos << cluster
         end