+++ CHANGES.md
@@ -23,6 +23,7 @@
 | s3 | S3バケット |
 | s3_regional | S3バケット(リージョン固有) |
 | vpc_peering | VPC Peering |
+| vpc_subnet | VPC Subnet |
 | wafv2 | AWS WAF(V2) |
 
 ## その他
@@ -47,12 +48,27 @@
 
 * yml設定でのリソース順を保持して出力
     * .furikake.yml で `keep_config_order: true` 指定
+* 取得対象のタグを指定(filtersでのtag指定に対応しているリソースのみ)
+    * .furikake.yml で `filters: []` 指定
+    * filters 対応リソースへの共通設定
+    * 末尾に備考として "抽出条件" を出力
+* 出力タグキーを指定(対応しているリソースのみ)
+    * .furikake.yml で `output_tag_keys: []` 指定
+    * 末尾に備考として "Tags列への出力タグキー" を出力
 
 ```yaml
-resources:
+options:
   keep_config_order: true
-  aws:
-    - ...
+  filters:
+    # tag キーと値での絞り込み例
+    - name: "tag:env"
+      values: ["dev", "stg"]
+    # tag キーのみでの絞り込み例
+    - name: "tag-key"
+      values: ["aws:cloudformation:stack-id", "aws:cloudformation:stack-name"]
+  output_tag_keys:
+    - "Name"
+    - "aws:autoscaling:groupName"
 ```
 
 ### Wiki作成(ProjectKey指定)